コメントを追加, 命名を改良
