+++ src/app/@sponsorSection/sponsor/sponsorAvatar.tsx
@@ -5,14 +5,16 @@ type SponsorAvatarProps = {
   icon: string;
 };
 
+// 画像の表示サイズはsponsorSectionが決定する
+// widthとheightは上位コンポーネントが制御するため、値は変更不要
 export default function SponsorAvatar({ icon }: SponsorAvatarProps) {
   return (
     <Image
       className="bg-white object-contain w-full aspect-[2/1] rounded-2xl mx-auto"
       alt="Icon"
       src={icon}
-      width={65} // 幅を指定
-      height={65} // 高さを指定
+      width={65}  
+      height={65}
     />
   );
 }
