+++ src/app/@sponsorSection/sponsor/sponsorType.tsx
@@ -1,12 +1,18 @@
 // consts/sponsor.ts
+
+// 個々のスポンサータイプの情報を定義
 export const Sponsor = {
+  // プラチナ
   PLATINUM: {
     typeName: "プラチナ",
+    // logosを変更することで、ロゴ画像の変更ができる
     logos: [
       "/images/Icon/icon_longTailedTit.png",
       "/images/Icon/icon_longTailedTit.png",
     ] as string[],
-    imageSize: "w-11/12 mx-auto",
+    // グリッドが表示される領域の横幅を指定
+    sponsorGridWidth: "w-11/12 mx-auto",
+    // レイアウトを指定
     layout: "grid gap-8",
   },
   LOCAL_GOLD: {
@@ -15,7 +21,7 @@ export const Sponsor = {
       "/images/Icon/icon_cantaloupe.png",
       "/images/Icon/icon_cantaloupe.png",
     ] as string[],
-    imageSize: "w-8/12 mx-auto",
+    sponsorGridWidth: "w-8/12 mx-auto",
     layout: "grid gap-8",
   },
   GOLD: {
@@ -24,7 +30,7 @@ export const Sponsor = {
       "/images/Icon/icon_lavender.png",
       "/images/Icon/icon_lavender.png",
     ] as string[],
-    imageSize: "w-8/12 mx-auto",
+    sponsorGridWidth: "w-8/12 mx-auto",
     layout: "grid gap-8",
   },
   LOCAL: {
@@ -33,19 +39,19 @@ export const Sponsor = {
       "/images/Icon/icon_lilyOfTheValley.png",
       "/images/Icon/icon_lilyOfTheValley.png",
     ] as string[],
-    imageSize: "w-8/12 mx-auto",
+    sponsorGridWidth: "w-8/12 mx-auto",
     layout: "grid gap-8",
   },
   DESIGN: {
     typeName: "デザイン",
     logos: ["/images/Icon/icon_owl.png"] as string[],
-    imageSize: "w-8/12 mx-auto",
+    sponsorGridWidth: "w-8/12 mx-auto",
     layout: "grid gap-8",
   },
   SPECIAL: {
     typeName: "スペシャル",
     logos: ["/images/Icon/icon_redCrownedCrane.png"] as string[],
-    imageSize: "w-8/12 mx-auto",
+    sponsorGridWidth: "w-8/12 mx-auto",
     layout: "grid gap-8",
   },
   SILVER: {
@@ -54,13 +60,13 @@ export const Sponsor = {
       "/images/Icon/icon_crab.png",
       "/images/Icon/icon_crab.png",
     ] as string[],
-    imageSize: "w-11/12 mx-auto",
+    sponsorGridWidth: "w-11/12 mx-auto",
     layout: "grid grid-cols-2 gap-8",
   },
 } as const;
 
+// 型安全にするための措置
 export type SponsorType = keyof typeof Sponsor;
-
 export type Sponsor = {
   type: SponsorType;
 };
@@ -70,9 +76,9 @@ export function getLogoList(sponsor: Sponsor): string[] {
   return Sponsor[sponsor.type].logos;
 }
 
-// スポンサータイプに応じてサイズクラスを返す
-export function getSponsorSize(sponsor: Sponsor): string {
-  return Sponsor[sponsor.type].imageSize;
+// スポンサータイプに応じてグリッドのサイズクラスを返す
+export function getSponsorGridWidth(sponsor: Sponsor): string {
+  return Sponsor[sponsor.type].sponsorGridWidth;
 }
 
 // スポンサータイプに応じてレイアウトクラスを返す
